+++ test/AuthorTest.php
@@ -2,7 +2,6 @@
 declare(strict_types=1);
 namespace Akismet;
 
-use function PHPUnit\Expect\{expect, it};
 use PHPUnit\Framework\{TestCase};
 
 /**
@@ -14,51 +13,46 @@ class AuthorTest extends TestCase {
    * @test Author::fromJson
    */
   public function testFromJson(): void {
-    it('should return a null reference with a non-object value', function() {
-      expect(Author::fromJson('foo'))->to->be->null;
-    });
+    // It should return a null reference with a non-object value.
+    assertThat(Author::fromJson('foo'), isNull());
 
-    it('should return an empty instance with an empty map', function() {
-      $author = Author::fromJson([]);
-      expect($author->getEmail())->to->be->empty;
-      expect($author->getIPAddress())->to->be->empty;
-    });
+    // It should return an empty instance with an empty map.
+    $author = Author::fromJson([]);
+    assertThat($author->getEmail(), isEmpty());
+    assertThat($author->getIPAddress(), isEmpty());
 
-    it('should return an initialized instance with a non-empty map', function() {
-      $author = Author::fromJson([
-        'comment_author_email' => 'cedric@belin.io',
-        'comment_author_url' => 'https://belin.io'
-      ]);
+    // It should return an initialized instance with a non-empty map.
+    $author = Author::fromJson([
+      'comment_author_email' => 'cedric@belin.io',
+      'comment_author_url' => 'https://belin.io'
+    ]);
 
-      expect($author->getEmail())->to->equal('cedric@belin.io');
-      expect((string) $author->getUrl())->to->equal('https://belin.io');
-    });
+    assertThat($author->getEmail(), equalTo('cedric@belin.io'));
+    assertThat((string) $author->getUrl(), equalTo('https://belin.io'));
   }
 
   /**
    * @test Author::jsonSerialize
    */
   public function testJsonSerialize(): void {
-    it('should return only the IP address and user agent with a newly created instance', function() {
-      $data = (new Author('127.0.0.1', 'Doom/6.6.6'))->jsonSerialize();
-      expect(get_object_vars($data))->to->have->lengthOf(2);
-      expect($data->user_agent)->to->equal('Doom/6.6.6');
-      expect($data->user_ip)->to->equal('127.0.0.1');
-    });
+    // It should return only the IP address and user agent with a newly created instance.
+    $data = (new Author('127.0.0.1', 'Doom/6.6.6'))->jsonSerialize();
+    assertThat(get_object_vars($data), countOf(2));
+    assertThat($data->user_agent, equalTo('Doom/6.6.6'));
+    assertThat($data->user_ip, equalTo('127.0.0.1'));
 
-    it('should return a non-empty map with a initialized instance', function() {
-      $data = (new Author('192.168.0.1', 'Mozilla/5.0', 'Cédric Belin'))
-        ->setEmail('cedric@belin.io')
-        ->setUrl('https://belin.io')
-        ->jsonSerialize();
+    // It should return a non-empty map with a initialized instance.
+    $data = (new Author('192.168.0.1', 'Mozilla/5.0', 'Cédric Belin'))
+      ->setEmail('cedric@belin.io')
+      ->setUrl('https://belin.io')
+      ->jsonSerialize();
 
-      expect(get_object_vars($data))->to->have->lengthOf(5);
-      expect($data->comment_author)->to->equal('Cédric Belin');
-      expect($data->comment_author_email)->to->equal('cedric@belin.io');
-      expect($data->comment_author_url)->to->equal('https://belin.io');
-      expect($data->user_agent)->to->equal('Mozilla/5.0');
-      expect($data->user_ip)->to->equal('192.168.0.1');
-    });
+    assertThat(get_object_vars($data), countOf(5));
+    assertThat($data->comment_author, equalTo('Cédric Belin'));
+    assertThat($data->comment_author_email, equalTo('cedric@belin.io'));
+    assertThat($data->comment_author_url, equalTo('https://belin.io'));
+    assertThat($data->user_agent, equalTo('Mozilla/5.0'));
+    assertThat($data->user_ip, equalTo('192.168.0.1'));
   }
 
   /**
@@ -69,16 +63,16 @@ class AuthorTest extends TestCase {
       ->setEmail('cedric@belin.io')
       ->setUrl('https://belin.io');
 
-    it('should start with the class name', function() use ($author) {
-      expect($author)->to->startWith('Akismet\Author {');
-    });
+    // It should start with the class name.
+    assertThat($author, stringStartsWith('Akismet\Author {'));
 
-    it('should contain the instance properties', function() use ($author) {
-      expect($author)->to->contain('"comment_author":"Cédric Belin"')
-        ->and->contain('"comment_author_email":"cedric@belin.io"')
-        ->and->contain('"comment_author_url":"https://belin.io"')
-        ->and->contain('"user_agent":"Doom/6.6.6"')
-        ->and->contain('"user_ip":"127.0.0.1"');
-    });
+    // It should contain the instance properties.
+    assertThat($author, logicalAnd(
+      stringContains('"comment_author":"Cédric Belin"'),
+      stringContains('"comment_author_email":"cedric@belin.io"'),
+      stringContains('"comment_author_url":"https://belin.io"'),
+      stringContains('"user_agent":"Doom/6.6.6"'),
+      stringContains('"user_ip":"127.0.0.1"')
+    ));
   }
 }